requirements.txt
aiosqlite==0.17.0
asgiref==3.3.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
bson==0.5.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
databases==0.4.1
Deprecated==1.2.11
Django==3.1.4
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
dnspython==2.1.0
docutils==0.16
drf-yasg==1.20.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-jwt-auth==0.5.0
fastapi-users==5.1.0
Flask==1.1.2
gunicorn==20.1.0
h11==0.9.0
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
lxml==4.6.3
makefun==1.9.5
Markdown==3.3.4
MarkupSafe==1.1.1
packaging==21.0
passlib==1.7.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.3
PyJWT==2.1.0
pymongo==3.11.3
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.5
pywin32==300
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
uvicorn==0.11.5
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
